+++ ontolib-cli/src/main/java/com/github/phenomics/ontolib/cli/Main.java
@@ -1,0 +1,45 @@
+package com.github.phenomics.ontolib.cli;
+
+import com.beust.jcommander.JCommander;
+import com.beust.jcommander.ParameterException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * OntoLib CLI implementation, main entry point.
+ *
+ * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
+ */
+public class Main {
+  /**
+   * {@link Logger} object to use.
+   */
+  private static final Logger LOGGER = LoggerFactory.getLogger(Main.class);
+
+  private static final String PRECOMPUTE_SCORES = "precompute-scores";
+
+  public static void main(String[] argv) {
+    final PrecomputeScoresOptions precomputeScoresOptions = new PrecomputeScoresOptions();
+    final JCommander jc =
+        JCommander.newBuilder().addCommand(PRECOMPUTE_SCORES, precomputeScoresOptions).build();
+    try {
+      jc.parse(argv);
+    } catch (ParameterException e) {
+      LOGGER.error("ERROR: " + e.getMessage());
+      jc.usage();
+      System.exit(1);
+    }
+
+    if (jc.getParsedCommand() == null) {
+      jc.usage();
+      System.exit(1);
+    }
+
+    switch (jc.getParsedCommand()) {
+      case PRECOMPUTE_SCORES:
+        new PrecomputeScoresCommand(precomputeScoresOptions).run();
+        break;
+    }
+  }
+
+}