+++ ontolib-cli/src/main/java/com/github/phenomics/ontolib/cli/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Command line interface to Ontolib.
+ *
+ * <p>
+ * Currently, this only gives access to score distribution precomputation.
+ * </p>
+ *
+ * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
+ */
+package com.github.phenomics.ontolib.cli;