+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 v0.2 (work in progress)
 -----------------------
 
+- Adding ``ontolib-cli`` package that allows score distribution precomputation from the command line.
 - Removed some dead code.
 - Added various tests, minor internal refactoring.
 - Moved ``OntologyTerms`` into ``ontology.algo`` package.